backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-mssql-backend==2.8.1
mssql-django==1.4
pyodbc==4.0.39
python-decouple==3.8
openpyxl==3.1.2

